change the way to install
