--- src/AnalyserNode/index.js
@@ -1,3 +1,0 @@
-"use strict";
-
-import "./getFloatTimeDomainData";
