--- src/AudioBuffer/index.js
@@ -1,4 +1,0 @@
-"use strict";
-
-import "./copyFromChannel";
-import "./copyToChannel";
