--- src/OfflineAudioContext/index.js
@@ -1,3 +1,0 @@
-"use strict";
-
-import "./startRendering";
